Former-commit-id: 3c1cdf6b2b8e129c1bef380f5b991af7e8aafd79

+++ src/Core/BoundedContexts/Supervisor/WB.Core.BoundedContexts.Supervisor/Views/DataExport/InterviewDataExportFactory.cs
@@ -150,7 +150,7 @@ namespace WB.Core.BoundedContexts.Supervisor.Views.DataExport
             dataRecords.Add(record);
         }
 
-        private int? GetParentRecordIndex(InterviewLevel level)
+        private decimal? GetParentRecordIndex(InterviewLevel level)
         {
             if (level.RosterVector.Length < 2)
                 return null;
