+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_answering_question_inside_propagatable_group_which_triggers_group_enablement_with_mandatory_question_inside_disabled_group.cs
@@ -77,7 +77,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
             
 
             interview = CreateInterview(questionnaireId: questionnaireId);
-            interview.AnswerNumericIntegerQuestion(userId, questionWhichIsForcesPropagationId, new int[] { }, DateTime.Now, 1);
+            interview.AnswerNumericIntegerQuestion(userId, questionWhichIsForcesPropagationId, new decimal[] { }, DateTime.Now, 1);
 
             expressionProcessor.Setup(x => x.EvaluateBooleanExpression(Moq.It.IsAny<string>(), Moq.It.IsAny<Func<string, object>>()))
                 .Returns(true);
@@ -86,7 +86,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         private Because of = () =>
-            interview.AnswerNumericRealQuestion(userId, answeringQuestionId, new int[] {0}, DateTime.Now, 0);
+            interview.AnswerNumericRealQuestion(userId, answeringQuestionId, new decimal[] { 0 }, DateTime.Now, 0);
 
         private Cleanup stuff = () =>
         {
